Modularize the code for neuro-integration Will need to clean it up tho
