+++ desktop/native/neuro-integration/go.mod
@@ -2,4 +2,6 @@ module github.com/neuro-desktop/integration-code
 
 go 1.22
 
-require github.com/gorilla/websocket v1.5.3 // indirect
+require (
+	github.com/gorilla/websocket v1.5.3 // indirect
+)
